requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.2.0
joblib>=1.2.0

# Computer vision & pose estimation
mediapipe>=0.10.0
opencv-python>=4.7.0
pillow>=9.0.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Jupyter / notebooks
ipykernel>=6.0.0
ipython>=8.0.0
jupyter_client>=7.0.0
jupyter_core>=5.0.0
jupyterlab>=4.0.0
nbformat>=5.7.0
nbconvert>=7.0.0

# Utilities
requests>=2.28.0
PyYAML>=6.0
tqdm>=4.64.0

# Deep Learning (Phase 3)
tensorflow>=2.10.0
imbalanced-learn>=0.10.0

# Web Deployment (Phase 4)
streamlit>=1.20.0

# MLOps (Phase 5)
mlflow>=2.0.0
